Revert parser extraction and OOP rebasing
Two steps forward, one step back
