+++ pyiron_continuum/fenics/job/elastic.py
@@ -16,7 +16,7 @@ with ImportAlarm(
 
 from pyiron_continuum.fenics.job.generic import Fenics
 from pyiron_continuum.fenics.plot import Plot
-from pyiron_continuum.fenics.factory import SolverConfig
+from pyiron_continuum.fenics.factory import SolverConfig, StringInputParser
 from pyiron_continuum.fenics.parser import FenicsString
 
 __author__ = "Liam Huber"
@@ -37,8 +37,8 @@ class SimpleBoundaries(HasStorage):
         self.storage.pairs = []
 
     def add(self, value, condition):
-        value = FenicsString(value)
-        condition = FenicsString(condition)
+        StringInputParser(value)
+        StringInputParser(condition)
         self.storage.pairs.append((value, condition))
 
     def list(self):
